re-factor and build migration to npm
