+++ DarkStatsCore.Data/Scraper.cs
@@ -1,12 +1,7 @@
 ﻿using System;
-using System.IO;
 using System.Net.Http;
 using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using System.Collections.Generic;
-using System.Collections.Concurrent;
-using System.Diagnostics;
 using HtmlAgilityPack;
 using Microsoft.EntityFrameworkCore;
 using DarkStatsCore.Data.Models;
@@ -16,17 +11,10 @@ namespace DarkStatsCore.Data
     public static class Scraper
     {
         public static List<long> Deltas = new List<long>();
-        public static DateTime LastGathered = DateTime.MinValue;
-        public static TimeSpan TimeSpanSinceLastCheck = TimeSpan.FromSeconds(0);
-        public static double ScrapeTimeAvg => _scrapeTime.Count() == 0 ? 0 : _scrapeTime.Average();
         public static EventHandler ScrapeSaved;
-        private static List<long> _scrapeTime = new List<long>();
-        private const int _scrapeTimeToKeep = 60;
         private static long _lastCheckTotalBytes = 0;
         private static int _deltasToKeep = 30;
         private static List<HostPadding> _hostPadding = new List<HostPadding>();        
-        private static Task _scrapeTask;
-        private static CancellationTokenSource _cancellationTokenSource;
 
         public static void Scrape(string url)
         {
@@ -43,7 +31,8 @@ namespace DarkStatsCore.Data
                                   .Where(t => t.Day.Month == DateTime.Now.Month && t.Day.Year == DateTime.Now.Year)
                                   .ToList();
             _hostPadding.RemoveAll(h => h.Month != DateTime.Now.Month && h.Year != DateTime.Now.Year);
-            if (stats.Sum(s => s.In + s.Out) + _hostPadding.Sum(h => h.In + h.Out) < traffic.Sum(t => t.In + t.Out))
+            long statsTotal = stats.Sum(s => s.In + s.Out);
+            if (statsTotal + _hostPadding.Sum(h => h.In + h.Out) < traffic.Sum(t => t.In + t.Out))
             {
                 Console.Write("Source stats are lower than db, calculating host padding... ");
                 _hostPadding = new List<HostPadding>();
@@ -82,9 +71,9 @@ namespace DarkStatsCore.Data
                 {
                     Deltas.RemoveAt(0);
                 }
-                Deltas.Add(stats.Sum(t => t.In + t.Out) - _lastCheckTotalBytes);
+                Deltas.Add(statsTotal - _lastCheckTotalBytes);
             }
-            _lastCheckTotalBytes = stats.Sum(t => t.In + t.Out);
+            _lastCheckTotalBytes = statsTotal;
 
             Console.Write("Saving... ");
             var lastRecords = context.TrafficStats
@@ -141,9 +130,6 @@ namespace DarkStatsCore.Data
             }
             context.SaveChanges();
 
-            var now = DateTime.Now;
-            TimeSpanSinceLastCheck = now.Subtract(LastGathered);
-            LastGathered = now;
             if (DashboardScrape.IsTaskActive)
             {
                 ScrapeSaved?.Invoke(null, EventArgs.Empty);
@@ -191,55 +177,6 @@ namespace DarkStatsCore.Data
             using (var httpClient = new HttpClient { Timeout = TimeSpan.FromMilliseconds(200) })
             { 
                 return httpClient.GetStringAsync(url + @"hosts/?full=yes&sort=total").Result;
-            }
-        }
-
-        public static void StartScrapeTask(TimeSpan saveTime, string url)
-        {
-            if (_scrapeTask != null)
-            {
-                _cancellationTokenSource.Cancel();
-            }
-            _cancellationTokenSource = new CancellationTokenSource();
-            _scrapeTask = ScrapeTask(url, saveTime, _cancellationTokenSource.Token);
-            DnsService.Start();
-        }
-        
-        private static async Task ScrapeTask(string url, TimeSpan saveTime, CancellationToken cancellationToken)
-        {
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                GatherData(url);
-                await Task.Delay(saveTime);
-            }
-        }
-
-        private static void GatherData(string url)
-        {
-            Console.Write(DateTime.Now.ToString("MM/dd/yyyy HH:mm:ss") + " - Gathering data... ");
-            var stopwatch = Stopwatch.StartNew();
-            try
-            {
-                Scraper.Scrape(url);
-                stopwatch.Stop();
-                Console.WriteLine("Done. (" + stopwatch.ElapsedMilliseconds + "ms)");
-                if (_scrapeTime.Count() == _scrapeTimeToKeep)
-                {
-                    _scrapeTime.RemoveAt(0);
-                }
-                _scrapeTime.Add(stopwatch.ElapsedMilliseconds);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine();
-                if (e.InnerException != null)
-                {
-                    Console.WriteLine(e.InnerException.Message + Environment.NewLine + e.InnerException.StackTrace);
-                }
-                else
-                {
-                    Console.WriteLine(e.Message + Environment.NewLine + e.StackTrace);
-                }
             }
         }
     }
