+++ DarkStatsCore/Pages/ViewDay.cshtml
@@ -4,7 +4,7 @@
 @{
     ViewData["Title"] = "Day";
 }
-<script src="~/lib/jquery-ui/jquery-ui.js"></script>
+<script src="~/lib/jquery-ui/jquery-ui.min.js"></script>
 <script>
   $( function() {
     $( "#datepicker" ).datepicker(
@@ -83,7 +83,7 @@
         <div class="col-sm-1"></div>
     </div>
 
-    <script src="~/lib/chartist/dist/chartist.min.js"></script>
+    <script src="~/lib/chartist/chartist.min.js"></script>
 
     <script>
         var data = {
