+++ DarkStatsCore/Pages/ViewMonth.cshtml
@@ -60,7 +60,7 @@
     </div>
 </div>
 
-<script src="~/lib/chartist/dist/chartist.min.js"></script>
+<script src="~/lib/chartist/chartist.min.js"></script>
 
 <script>
     var data = {
