+++ DarkStatsCore/Pages/_Layout.cshtml
@@ -4,15 +4,15 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>@ViewData["Title"] - DarkStatsCore</title>
-    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.css" />
+    <link rel="stylesheet" href="~/lib/bootstrap/bootstrap.min.css" />
     <link rel="stylesheet" href="~/css/site.css" />
-	<link rel="stylesheet" href="~/lib/chartist/dist/chartist.min.css">
+	<link rel="stylesheet" href="~/lib/chartist/chartist.min.css">
     <link rel="stylesheet" href="~/lib/jquery-ui/themes/base/jquery-ui.min.css">
 </head>
 <body>
 
-    <script src="~/lib/jquery/dist/jquery.js"></script>
-    <script src="~/lib/bootstrap/dist/js/bootstrap.js"></script>
+    <script src="~/lib/jquery/jquery.min.js"></script>
+    <script src="~/lib/bootstrap/bootstrap.min.js"></script>
     <script src="~/js/site.js" asp-append-version="true"></script>
 
     @RenderSection("Scripts", required: false)
